pyproject.toml
[tool.poetry]
name = "advanced-machine-learning"
version = "0.1.0"
description = "Human-Machine boundary detection with DeBERTaV3 and CRF"
authors = ["dlongobardi@unisa.it"]
packages = [{include = "src"}]

[tool.poetry.dependencies]
python = ">=3.10,<3.15"  
torch = "*"
transformers = "*"
hydra-core = "*"
mlflow = "*"
pandas = "*"
openpyxl = "*"  # Necessario per leggere il file raw_tribert.xlsx
pytorch-crf = "*"
tiktoken = "^0.12.0"
protobuf = "<=3.20.3"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"